feature: add oam model
