+++ scaleph-application/scaleph-application-oam/src/main/java/cn/sliew/scaleph/application/oam/model/application/ApplicationComponent.java
@@ -16,14 +16,20 @@
  * limitations under the License.
  */
 
-package cn.sliew.scaleph.kubernetes.oam.common;
+package cn.sliew.scaleph.application.oam.model.application;
 
 import lombok.Data;
 
-@Data
-public class AbstractSchema implements Schema {
+import java.util.List;
+import java.util.Properties;
 
-    private String apiVersion;
-    private String kind;
-    private MetaData metadata;
+@Data
+public class ApplicationComponent {
+
+    private String name;
+    private String type;
+    private Properties properties;
+    private Object inputs;
+    private Object outputs;
+    private List<ApplicationComponentTrait> traits;
 }
