+++ scaleph-application/scaleph-application-oam/src/main/java/cn/sliew/scaleph/application/oam/model/application/ApplicationConfiguration.java
@@ -16,9 +16,9 @@
  * limitations under the License.
  */
 
-package cn.sliew.scaleph.kubernetes.oam.definition;
+package cn.sliew.scaleph.application.oam.model.application;
 
-import cn.sliew.scaleph.kubernetes.oam.common.AbstractSchema;
+import cn.sliew.scaleph.application.oam.model.common.AbstractSchema;
 import io.fabric8.kubernetes.model.annotation.Group;
 import io.fabric8.kubernetes.model.annotation.Version;
 import lombok.Data;
@@ -26,41 +26,6 @@ import lombok.Data;
 @Data
 @Group("core.oam.dev")
 @Version("v1beta1")
-public class ComponentDefinition extends AbstractSchema {
+public class ApplicationConfiguration extends AbstractSchema {
 
-    private Spec spec;
-
-    @Data
-    public static class Spec {
-
-        private Semantic semantic;
-        private WorkloadTypeDescriptor workload;
-    }
-
-    @Data
-    public static class Semantic {
-        // kubevela 支持 cue、helm、kube
-    }
-
-    @Data
-    public static class WorkloadTypeDescriptor {
-
-        /**
-         * 通过 name 引用 WorkloadDefinition
-         */
-        private String type;
-
-        /**
-         * 通过 group、version 和 kind 引用 WorkloadDefinition。gvk -> group、version、kind
-         * 和 type 互斥，只能同时存在一个。
-         */
-        private WorkloadGVK definition;
-    }
-
-    @Data
-    public static class WorkloadGVK {
-
-        private String apiVersion;
-        private String kind;
-    }
 }
