+++ scaleph-application/scaleph-application-oam/src/main/java/cn/sliew/scaleph/application/oam/model/common/ComponentBinder.java
@@ -16,17 +16,7 @@
  * limitations under the License.
  */
 
-package cn.sliew.scaleph.kubernetes.oam.common;
+package cn.sliew.scaleph.application.oam.model.common;
 
-import lombok.Data;
-
-@Data
-public class DefinitionRef {
-
-    /**
-     * name 和 apiVersion、kind 互斥，只能使用一种方式
-     */
-    private String name;
-    private String apiVersion;
-    private String kind;
+public interface ComponentBinder {
 }
