+++ scaleph-application/scaleph-application-oam/src/main/java/cn/sliew/scaleph/application/oam/model/common/MetaData.java
@@ -16,24 +16,21 @@
  * limitations under the License.
  */
 
-package cn.sliew.scaleph.kubernetes.oam.definition;
+package cn.sliew.scaleph.application.oam.model.common;
 
-import cn.sliew.scaleph.kubernetes.oam.common.AbstractSchema;
-import cn.sliew.scaleph.kubernetes.oam.common.DefinitionRef;
-import io.fabric8.kubernetes.model.annotation.Group;
-import io.fabric8.kubernetes.model.annotation.Version;
 import lombok.Data;
 
+import java.util.Map;
+
 @Data
-@Group("core.oam.dev")
-@Version("v1beta1")
-public class WorkloadDefinition extends AbstractSchema {
+public class MetaData {
 
-    private Spec spec;
-
-    @Data
-    public static class Spec {
-
-        private DefinitionRef definitionRef;
-    }
+    private String name;
+    private Map<String, String> labels;
+    /**
+     * version 和 description 预定义且推荐使用
+     * definition.oam.dev/description: description
+     * definition.oam.dev/version: version
+     */
+    private Map<String, String> annotations;
 }
