+++ scaleph-application/scaleph-application-oam/src/main/java/cn/sliew/scaleph/application/oam/model/common/Schema.java
@@ -16,16 +16,7 @@
  * limitations under the License.
  */
 
-package cn.sliew.scaleph.kubernetes.oam.application;
+package cn.sliew.scaleph.application.oam.model.common;
 
-import cn.sliew.scaleph.kubernetes.oam.common.AbstractSchema;
-import io.fabric8.kubernetes.model.annotation.Group;
-import io.fabric8.kubernetes.model.annotation.Version;
-import lombok.Data;
-
-@Data
-@Group("core.oam.dev")
-@Version("v1beta1")
-public class ApplicationConfiguration extends AbstractSchema {
-
+public interface Schema {
 }
