+++ scaleph-application/scaleph-application-oam/src/main/java/cn/sliew/scaleph/application/oam/model/common/Schematic.java
@@ -16,28 +16,13 @@
  * limitations under the License.
  */
 
-package cn.sliew.scaleph.kubernetes.oam.definition;
+package cn.sliew.scaleph.application.oam.model.common;
 
-import cn.sliew.scaleph.kubernetes.oam.common.AbstractSchema;
-import io.fabric8.kubernetes.model.annotation.Group;
-import io.fabric8.kubernetes.model.annotation.Version;
 import lombok.Data;
 
 @Data
-@Group("core.oam.dev")
-@Version("v1beta1")
-public class WorkflowStepDefinition extends AbstractSchema {
+public class Schematic {
 
-    private Spec spec;
-
-    @Data
-    public static class Spec {
-        private Semantic schematic;
-    }
-
-    @Data
-    public static class Semantic {
-        // kubevela 支持 cue、helm、kube
-    }
-
+    // kubevela 支持 cue、helm、kube
+    private String cue;
 }
