+++ scaleph-application/scaleph-application-oam/src/main/java/cn/sliew/scaleph/application/oam/model/common/WorkloadBinder.java
@@ -16,25 +16,14 @@
  * limitations under the License.
  */
 
-package cn.sliew.scaleph.kubernetes.oam.definition;
+package cn.sliew.scaleph.application.oam.model.common;
 
-import cn.sliew.scaleph.kubernetes.oam.common.AbstractSchema;
-import cn.sliew.scaleph.kubernetes.oam.common.DefinitionRef;
-import io.fabric8.kubernetes.model.annotation.Group;
-import io.fabric8.kubernetes.model.annotation.Version;
-import lombok.Data;
+import cn.sliew.scaleph.application.oam.model.definition.ComponentDefinition;
+import cn.sliew.scaleph.application.oam.model.definition.WorkloadDefinition;
 
-@Deprecated
-@Data
-@Group("core.oam.dev")
-@Version("v1beta1")
-public class ScopeDefinition extends AbstractSchema {
+public interface WorkloadBinder {
 
-    private Spec spec;
+    WorkloadDefinition getWorkload();
 
-    @Data
-    public static class Spec {
-
-        private DefinitionRef definitionRef;
-    }
+    ComponentDefinition getComponent();
 }
