+++ scaleph-application/scaleph-application-oam/src/main/java/cn/sliew/scaleph/application/oam/model/definition/PolicyDefinition.java
@@ -16,7 +16,25 @@
  * limitations under the License.
  */
 
-package cn.sliew.scaleph.kubernetes.oam.common;
+package cn.sliew.scaleph.application.oam.model.definition;
 
-public interface ComponentBinder {
+import cn.sliew.scaleph.application.oam.model.common.Schematic;
+import cn.sliew.scaleph.application.oam.model.common.AbstractSchema;
+import cn.sliew.scaleph.application.oam.model.common.DefinitionRef;
+import io.fabric8.kubernetes.model.annotation.Group;
+import io.fabric8.kubernetes.model.annotation.Version;
+import lombok.Data;
+
+@Data
+@Group("core.oam.dev")
+@Version("v1beta1")
+public class PolicyDefinition extends AbstractSchema {
+
+    private Spec spec;
+
+    @Data
+    public static class Spec {
+        private DefinitionRef definitionRef;
+        private Schematic schematic;
+    }
 }
