+++ scaleph-dao/src/main/java/cn/sliew/scaleph/dao/entity/master/oam/OamTraitDefinition.java
@@ -21,7 +21,6 @@ package cn.sliew.scaleph.dao.entity.master.oam;
 import cn.sliew.scaleph.dao.entity.BaseDO;
 import com.baomidou.mybatisplus.annotation.TableField;
 import com.baomidou.mybatisplus.annotation.TableName;
-import io.swagger.v3.oas.annotations.media.Schema;
 import lombok.Data;
 
 /**
@@ -29,12 +28,10 @@ import lombok.Data;
  */
 @Data
 @TableName("oam_trait_definition")
-@Schema(name = "OamTraitDefinition", description = "Trait Definition 信息")
 public class OamTraitDefinition extends BaseDO {
 
     private static final long serialVersionUID = 1L;
 
-    @Schema(description = "定义id。主要用于 kubernetes 中 metadata 使用")
     @TableField("definition_id")
     private String definitionId;
 
@@ -53,7 +50,6 @@ public class OamTraitDefinition extends BaseDO {
     @TableField("properties")
     private String properties;
 
-    @Schema(description = "备注")
     @TableField("remark")
     private String remark;
 }
